Routing and stream opterations cleaned
