+++ .gitignore
@@ -1,6 +1,6 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-/streams/api/**
+streams/api/**
 # dependencies
 /node_modules
 /.pnp
