+++ streams/api/db.json
@@ -1,51 +1,3 @@
 {
-  "streams": [
-    {
-      "title": "A Stream",
-      "description": "A River",
-      "userId": "115752212371876410838",
-      "id": 1
-    },
-    {
-      "title": "This is a stream",
-      "description": "It seems to be working",
-      "userId": "115752212371876410838",
-      "id": 2
-    },
-    {
-      "title": "Just Kidding!",
-      "description": "It's not working",
-      "userId": "115752212371876410838",
-      "id": 3
-    },
-    {
-      "title": "FuckFuck",
-      "description": "this edit was a PATCH request",
-      "userId": "115752212371876410838",
-      "id": 4
-    },
-    {
-      "title": "I think its actually working",
-      "description": "A STREAM thats edited",
-      "id": 5
-    },
-    {
-      "title": "If brooks babble what do streams do ?",
-      "description": "flow ? rush ? roar ? splash ?",
-      "userId": "115752212371876410838",
-      "id": 6
-    },
-    {
-      "title": "Is this working",
-      "description": "I think so!",
-      "userId": "115752212371876410838",
-      "id": 7
-    },
-    {
-      "title": "Stream Component Reuse!",
-      "description": "An example of component reuse changed",
-      "userId": "115752212371876410838",
-      "id": 8
-    }
-  ]
+  "streams": [ ]
 }
