+++ streams/client/src/components/App.js
@@ -16,7 +16,7 @@ const App = () => {
           <Header />
           <Route path="/" exact component={StreamList} />
           <Route path="/streams/new" exact component={StreamCreate} />
-          <Route path="/streams/edit" exact component={StreamEdit} />
+          <Route path="/streams/edit/:id" exact component={StreamEdit} />
           <Route path="/streams/delete" exact component={StreamDelete} />
           <Route path="/streams/show" exact component={StreamShow} />
         </div>
