+++ streams/client/src/components/streams/StreamEdit.js
@@ -1,7 +1,43 @@
+/* eslint-disable react/destructuring-assignment */
+import _ from 'lodash';
 import React from 'react';
+import { connect } from 'react-redux';
+import { formValues } from 'redux-form';
+import { fetchStream, editStream } from '../../actions';
+import StreamForm from './StreamForm';
 
-const StreamEdit = () => {
-  return <div>StreamEdit</div>;
+class StreamEdit extends React.Component {
+  componentDidMount() {
+    this.props.fetchStream(this.props.match.params.id);
+  }
+
+  onSubmit = (formValues) => {
+    this.props.editStream(this.props.match.params.id, formValues);
+  };
+
+  render() {
+    if (!this.props.stream) {
+      return <div>Loading...</div>;
+    }
+
+    // initialValues is special form prop which we have access to because StreamForm
+    // has a redux-form wrapper on it
+    // .pick is just creating an object with the specified properties
+    return (
+      <div>
+        <h3>Edit a Stream</h3>
+        <StreamForm
+          initialValues={_.pick(this.props.stream, 'title', 'description')}
+          onSubmit={this.onSubmit}
+        />
+      </div>
+    );
+  }
+}
+
+const mapStateToProps = (state, ownProps) => {
+  return { stream: state.streams[ownProps.match.params.id] };
 };
-
-export default StreamEdit;
+export default connect(mapStateToProps, { fetchStream, editStream })(
+  StreamEdit
+);
